# CPM-101: use symfony cache instead of doctrine cache