--- a/composer.json
+++ b/composer.json
@@ -54,6 +54,7 @@
         "ass/xmlsecurity": "1.1.1",
         "box/spout": "2.7.3",
         "doctrine/annotations": "1.8.0",
+        "doctrine/cache": "1.9.1",
         "doctrine/collections": "1.6.4",
         "doctrine/common": "2.11.0",
         "doctrine/data-fixtures": "1.4.0",