--- a/config/packages/doctrine.yml
+++ b/config/packages/doctrine.yml
@@ -67,4 +67,3 @@ doctrine_migrations:
     dir_name: '%doctrine_migrations_dir%'
     namespace: Pim\Upgrade\Schema
     name: PIM Migrations
-    column_length: 255