--- a/symfony.lock
+++ b/symfony.lock
@@ -424,6 +424,9 @@
     "psr/event-dispatcher": {
         "version": "1.0.0"
     },
+    "psr/http-client": {
+        "version": "1.0.1"
+    },
     "psr/http-factory": {
         "version": "1.0.1"
     },