Doctrine cache bundle is no longer maintained: doctrine/DoctrineCacheBundle#156

Since symfony 4.3 the best way to configure the cache is using the `symfony/cache` component. In this PR the doctrine-cache-bundle is removed and replaced by symfony/cache. The config is migrated to this new component.  


**Which configuration to use?**

In our previous configuration we had 

```yaml
doctrine_cache:
    providers:
        pim_apcu:
            apcu: ~
        pim_array:
            array: ~
        pim:
            chain:
                providers: [doctrine_cache.providers.pim_apcu, doctrine_cache.providers.pim_array]
```



The default configuration in SF5 install is defined here: https://github.com/symfony/recipes/blob/master/doctrine/doctrine-bundle/1.6/config/packages/prod/doctrine.yaml  
It seems simpler compared to our previous config with a simple and obscure `cache.app`:

```yaml
framework:
    cache:
        pools:
            doctrine.result_cache_pool:
                adapter: cache.app
            doctrine.system_cache_pool:
                adapter: cache.system
```

So I adapt the symfony cache configuration to have the same behavior than before for results only:

```yaml
framework:
    cache:
        pools:
            doctrine.result_cache_pool:
                adapters:
                    - cache.adapter.apcu
                    - cache.adapter.array
            doctrine.system_cache_pool:
                adapter: cache.system
```

The `doctrine.system_cache_pool` is different and used the default `cache.system` cache. Here is an explaination of this cache https://symfony.com/doc/current/cache.html#configuring-cache-with-frameworkbundle


**Benchmarks in local**

Load and force save 1239 products 5 times in the same process without clearing cache. The results are the same. The cost to fetch/save product from/in the cache is really low compared to the rest.

| Duration in seconds | master | this PR |
|---------------------|--------|---------|
| Pass 1              | 127    | 124     |
| Pass 2              | 124    | 133     |
| Pass 3              | 125    | 132     |
| Pass 4              | 120    | 118     |
| Pass 5              | 126    | 112     |
| | | |
| Average             | 124.4  | 123.8   |
| Median              | 125    | 124     |